refactor: Move Composable from CourseUI to ComposeCommon
Fixes: LEARNER-9716
