+++ course/src/main/res/values/strings.xml
@@ -34,7 +34,6 @@
     <string name="course_you_cant_enroll">You cannot enroll in this course because the enrollment date is over.</string>
     <string name="course_not_started">This course hasn’t started yet.</string>
     <string name="course_access_expired">Your access to this course has expired.</string>
-    <string name="course_not_connected_to_internet">You are not connected to the Internet. Please check your Internet connection.</string>
     <string name="course_navigation_course">Course</string>
     <string name="course_navigation_video">Video</string>
     <string name="course_navigation_discussion">Discussion</string>
@@ -49,6 +48,7 @@
     <string name="course_to_proceed">To proceed with \"%s\" press \"Next section\".</string>
     <string name="course_gated_content_label">Some content in this part of the course is locked for upgraded users only.</string>
     <string name="course_navigation_dates" tools:ignore="MissingTranslation">Dates</string>
+    <string name="course_you_are_already_enrolled" tools:ignore="MissingTranslation">You are already enrolled in this course.</string>
 
     <!-- Error Message when course dates are not available -->
     <string name="course_dates_unavailable_message" tools:ignore="MissingTranslation">Course dates are not currently available.</string>
